// src/lib/model/Clip.ts
export type Clip = {
  index: number
  clipId: string
  clipUrl: string
  clipEmbedUrl: string
  clipTitle: string
  clipThumbnailUrl: string
  clipCreatedAt: string
  broadcasterId: string
  broadcasterName: string
  creatorId: string
  creatorName: string
  videoId: string
  vodOffset: number
  videoCreatedAt: string
}